add SharedPreferencesUtil class
