+++ app/src/main/AndroidManifest.xml
@@ -77,6 +77,7 @@
 
         <activity android:name=".ui.splash.SplashActivity" />
         <activity android:name=".ui.register.RegisterActivity" />
+        <activity android:name=".ui.share.ShareActivity" />
         <activity
             android:name=".ui.search.SearchActivity"
             android:noHistory="true" />
