+++ app/src/main/java/vn/asiantech/way/ui/base/BaseActivity.kt
@@ -1,7 +1,5 @@
 package vn.asiantech.way.ui.base
 
-import android.content.Context
-import android.content.SharedPreferences
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 import io.reactivex.disposables.CompositeDisposable
@@ -12,14 +10,8 @@ import io.reactivex.disposables.Disposable
  * Created by quocnguyenp. on 9/21/17.
  */
 abstract class BaseActivity : AppCompatActivity() {
-    companion object {
-        const val SHARED_NAME = "shared"
-        const val KEY_LOGIN = "login"
-        const val KEY_LOCAL_USER = "local_user"
-    }
 
     private val subscription: CompositeDisposable = CompositeDisposable()
-    protected lateinit var prefs: SharedPreferences
 
     protected fun addDisposables(vararg ds: Disposable) {
         ds.forEach { subscription.add(it) }
@@ -37,7 +29,6 @@ abstract class BaseActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        prefs = getSharedPreferences(SHARED_NAME, Context.MODE_PRIVATE)
     }
 
     /**
