+++ app/src/main/java/vn/asiantech/way/ui/search/SearchActivity.kt
@@ -3,12 +3,9 @@ package vn.asiantech.way.ui.search
 import android.app.ProgressDialog
 import android.content.Intent
 import android.os.Bundle
-import com.google.gson.Gson
-import com.google.gson.JsonSyntaxException
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.subjects.PublishSubject
 import org.jetbrains.anko.setContentView
-import org.json.JSONArray
 import vn.asiantech.way.R
 import vn.asiantech.way.data.model.AutoCompleteLocation
 import vn.asiantech.way.data.model.WayLocation
@@ -17,6 +14,7 @@ import vn.asiantech.way.extension.toast
 import vn.asiantech.way.ui.base.BaseActivity
 import vn.asiantech.way.ui.share.ShareActivity
 import vn.asiantech.way.utils.AppConstants
+import vn.asiantech.way.utils.SharedPreferencesUtil
 import java.util.concurrent.TimeUnit
 
 /**
@@ -31,11 +29,13 @@ class SearchActivity : BaseActivity() {
     private val searchObservable = PublishSubject.create<String>()
     private lateinit var startShareActivityIntent: Intent
     private lateinit var progressDialog: ProgressDialog
+    private lateinit var prefs: SharedPreferencesUtil
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         ui = SearchActivityUI(locations)
         ui.setContentView(this)
+        prefs = SharedPreferencesUtil(this)
         startShareActivityIntent = Intent(this, ShareActivity::class.java)
         progressDialog = ProgressDialog(this)
         progressDialog.setCancelable(false)
@@ -104,7 +104,7 @@ class SearchActivity : BaseActivity() {
      */
     internal fun loadSearchHistory() {
         locations.clear()
-        val histories = getSearchHistory()
+        val histories = prefs.getSearchHistory()
         if (histories != null) {
             locations.addAll(histories)
         }
@@ -129,45 +129,8 @@ class SearchActivity : BaseActivity() {
         }
     }
 
-    private fun getSearchHistory(): List<WayLocation>? {
-        val gson = Gson()
-        val result = mutableListOf<WayLocation>()
-        return try {
-            val history = prefs.getString(AppConstants.KEY_SEARCH_SCREEN_WAY_LOCATION_HISTORY,
-                    "[]")
-            val jsonArray = JSONArray(history)
-            (0 until jsonArray.length())
-                    .mapTo(result) {
-                        gson.fromJson(jsonArray.getJSONObject(it).toString()
-                                , WayLocation::class.java)
-                    }
-            result.toList()
-        } catch (e: JsonSyntaxException) {
-            null
-        }
-    }
-
-    private fun saveSearchHistory(location: WayLocation) {
-        val gson = Gson()
-        val editor = prefs.edit()
-        var history = getSearchHistory()
-        if (history == null) {
-            history = mutableListOf()
-        }
-        history = history.filter {
-            it.id != location.id
-        }.toMutableList()
-        location.isHistory = true
-        history.add(0, location)
-        if (history.size > AppConstants.SEARCH_SCREEN_HISTORY_MAX_SIZE) {
-            history.removeAt(AppConstants.SEARCH_SCREEN_HISTORY_MAX_SIZE - 1)
-        }
-        editor?.putString(AppConstants.KEY_SEARCH_SCREEN_WAY_LOCATION_HISTORY, gson.toJson(history))
-        editor?.apply()
-    }
-
     private fun startSharedActivity(location: WayLocation) {
-        saveSearchHistory(location)
+        prefs.saveSearchHistory(location)
         val bundle = Bundle()
         bundle.putParcelable(AppConstants.KEY_LOCATION, location)
         bundle.putString(AppConstants.KEY_CONFIRM,
